Fetch transformation plan
