+++ app/javascript/react/screens/App/Plan/PlanActions.js
@@ -1,9 +1,10 @@
 import URI from 'urijs';
 import API, { globalMockMode } from '../../../../common/API';
 
-import { FETCH_V2V_PLAN_REQUEST } from './PlanConstants';
+import { FETCH_V2V_PLAN_REQUEST, FETCH_V2V_PLAN } from './PlanConstants';
 
-import { requestPlanRequestData } from './plan.fixtures';
+import { requestPlanData } from './plan.fixtures';
+import { requestPlanRequestData } from './plan.planRequests.fixtures';
 
 const mockMode = globalMockMode;
 
@@ -25,4 +26,25 @@ const _getPlanRequestActionCreator = url => dispatch => {
 export const fetchPlanRequestAction = url => {
   const uri = new URI(url);
   return _getPlanRequestActionCreator(uri.toString());
+};
+
+// *****************************************************************************
+// * FETCH_V2V_PLAN
+// *****************************************************************************
+export const _getPlanActionCreator = (url, id) => dispatch => {
+  if (mockMode) {
+    return dispatch({
+      type: `${FETCH_V2V_PLAN}_FULFILLED`,
+      payload: requestPlanData(id).response
+    });
+  }
+  return dispatch({
+    type: FETCH_V2V_PLAN,
+    payload: API.get(url)
+  });
+};
+
+export const fetchPlanAction = (urlBuilder, id) => {
+  const uri = new URI(urlBuilder(id));
+  return _getPlanActionCreator(uri.toString(), id);
 };
