+++ app/javascript/react/screens/App/Plan/PlanConstants.js
@@ -1,1 +1,2 @@
 export const FETCH_V2V_PLAN_REQUEST = 'FETCH_V2V_PLAN_REQUEST';
+export const FETCH_V2V_PLAN = 'FETCH_V2V_PLAN';
